# back_end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MusicApp API",
        "description": "Documentation de l'API MusicApp (auth, artistes, albums, chansons.)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/show-artists": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Get all artists",
                "description": "Retrieve a list of all artists",
                "operationId": "92bc42919bd3bf67dadcb3b8b37b3a0e",
                "responses": {
                    "200": {
                        "description": "List of artists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Cheb Khaled"
                                            },
                                            "genre": {
                                                "type": "string",
                                                "example": "Raï"
                                            },
                                            "pays": {
                                                "type": "string",
                                                "example": "Algérie"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/artists": {
            "post": {
                "tags": [
                    "Artists"
                ],
                "summary": "Create a new artist",
                "description": "Store a newly created artist in the database",
                "operationId": "ace612d64eb15e36b094b94b7c30110b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "genre",
                                    "pays"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Cheb Khaled"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "example": "Raï"
                                    },
                                    "pays": {
                                        "type": "string",
                                        "example": "Algérie"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Artist created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Cheb Khaled"
                                        },
                                        "genre": {
                                            "type": "string",
                                            "example": "Raï"
                                        },
                                        "pays": {
                                            "type": "string",
                                            "example": "Algérie"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/artists/{id}": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Get artist details",
                "description": "Retrieve the details of a specific artist by ID",
                "operationId": "9f33c9ca7247e2b98ade31174cb981f9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Artist ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artist details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Artist name"
                                        },
                                        "genre": {
                                            "type": "string",
                                            "example": "Artist genre "
                                        },
                                        "pays": {
                                            "type": "string",
                                            "example": "Artist pays"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Artist not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Artists"
                ],
                "summary": "Update an artist",
                "description": "Update the details of an existing artist by ID",
                "operationId": "e3769486421ccb8d1a0f49ac6e72a9e2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Artist ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Name"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "example": "Updated Genre"
                                    },
                                    "pays": {
                                        "type": "string",
                                        "example": "Maroc"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Artist updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Updated Name"
                                        },
                                        "genre": {
                                            "type": "string",
                                            "example": "Updated Genre"
                                        },
                                        "pays": {
                                            "type": "string",
                                            "example": "Maroc"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Artist not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Artists"
                ],
                "summary": "Delete an artist",
                "description": "Delete an existing artist by ID",
                "operationId": "73ce1f07ca3ded736b444e8590fd055e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Artist ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artist deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Artist deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Artist not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use your API token from /login",
                "bearerFormat": "API Token",
                "scheme": "bearer"
            }
        }
    },
...
}